# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Peluquería SaaS API",
    "version": "1.0.0",
    "description": "API completa para gestión de peluquerías"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de desarrollo"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "email", "role", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["superadmin", "admin", "barber", "client"]
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Organization": {
        "type": "object",
        "required": ["id", "name", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "admin_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Barber": {
        "type": "object",
        "required": ["id", "user_id", "organization_id", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "schedule": {
            "type": "object",
            "properties": {
              "monday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "tuesday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "wednesday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "thursday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "friday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "saturday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              },
              "sunday": {
                "type": "object",
                "properties": {
                  "start": {"type": "string", "format": "time"},
                  "end": {"type": "string", "format": "time"},
                  "available": {"type": "boolean"}
                }
              }
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Client": {
        "type": "object",
        "required": ["id", "user_id", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "preferences": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Appointment": {
        "type": "object",
        "required": ["id", "barber_id", "client_id", "appointment_date", "status", "created_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "barber_id": {
            "type": "string",
            "format": "uuid"
          },
          "client_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "appointment_date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "default": 30
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "completed", "cancelled"]
          },
          "notes": {
            "type": "string"
          },
          "google_event_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "barber": {
            "$ref": "#/components/schemas/Barber"
          },
          "client": {
            "$ref": "#/components/schemas/Client"
          }
        }
      },
...
      "CreateOrganizationRequest": {
        "type": "object",
        "required": ["name", "admin_email", "admin_password", "admin_first_name", "admin_last_name"],
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "admin_email": {
            "type": "string",
            "format": "email"
          },
          "admin_password": {
            "type": "string",
            "minLength": 6
          },
          "admin_first_name": {
            "type": "string"
          },
          "admin_last_name": {
            "type": "string"
          }
        }
      },
      "CreateAdminRequest": {
        "type": "object",
        "required": ["organization_id", "email", "password", "first_name", "last_name"],
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
...
      "CreateAppointmentRequest": {
        "type": "object",
        "required": ["barber_id", "appointment_date"],
        "properties": {
          "barber_id": {
            "type": "string",
            "format": "uuid"
          },
          "appointment_date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "default": 30
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "UpdateAppointmentRequest": {
        "type": "object",
        "properties": {
          "appointment_date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "confirmed", "completed", "cancelled"]
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/superadmin/organizations": {
      "post": {
        "tags": ["SuperAdmin"],
        "summary": "Crear organización",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organización creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["SuperAdmin"],
        "summary": "Listar organizaciones",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Lista de organizaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/superadmin/admins": {
      "post": {
        "tags": ["SuperAdmin"],
        "summary": "Crear admin",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/superadmin/users": {
      "get": {
        "tags": ["SuperAdmin"],
        "summary": "Listar usuarios",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/appointments": {
      "post": {
        "tags": ["Appointments"],
        "summary": "Crear cita",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cita creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Appointments"],
        "summary": "Listar citas",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de citas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/barbers/{id}/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Listar citas de barbero",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de citas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clients/{id}/appointments": {
      "get": {
        "tags": ["Appointments"],
        "summary": "Listar citas de cliente",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de citas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Appointment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}": {
      "patch": {
        "tags": ["Appointments"],
        "summary": "Actualizar cita",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cita actualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Appointments"],
        "summary": "Eliminar cita",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cita eliminada"
          }
        }
      }
    },
...
    "/google/sync/{appointment_id}": {
      "post": {
        "tags": ["Google"],
        "summary": "Sincronizar cita con Google Calendar",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cita sincronizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Appointment"
                }
              }
            }
          }
        }
      }
    },
    "/google/event/{appointment_id}": {
      "delete": {
        "tags": ["Google"],
        "summary": "Eliminar evento de Google Calendar",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento eliminado"
          }
        }
      }
    }
  }
}